requirements.txt
opencv-python
pytesseract
Pillow
numpy
cvzone
python-barcode